+++ Services/WebApp/src/routes/api/system/[id]/reject/+server.ts
@@ -1,25 +1,22 @@
 import { systems, type System } from '$lib/api/systems';
+import type { User } from '$lib/database/userDatabase';
+import { error, json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
 
-export const GET: RequestHandler = async ({params}) => {
-    const system: System | undefined = systems.find((system) => system.id === params.id);
+export const GET: RequestHandler = async ({ params, fetch }) => {
+	const user: User | null = await fetch('/api/auth?admin=true')
+		.then((res) => res.json())
+		.catch(() => null);
 
-    if (!system) {
-        return new Response(
-            JSON.stringify({error: 'Invalid ID'}),
-            {
-                headers: {
-                    'content-type': 'application/json'
-                }
-            }
-        );
-    }
+	if (!user) throw error(401, 'Unauthorized');
 
-    system.status = 'Rejected';
+	const system: System | undefined = systems.find((system) => system.id === params.id);
 
-    return new Response(JSON.stringify(system), {
-        headers: {
-            'content-type': 'application/json'
-        }
-    });
+	if (!system) {
+		return json({ error: 'Invalid ID' });
+	}
+
+	system.status = 'Rejected';
+
+	return json(system);
 };
