+++ Services/WebApp/src/routes/hook/extraction/[provider_slug]/callback/+server.ts
@@ -3,7 +3,9 @@ import type { RequestHandler } from './$types';
 import type { Collection, Db, MongoClient, WithId } from 'mongodb';
 import clientPromise from '$lib/database/clientPromise';
 
-export const GET: RequestHandler = async ({ params }) => {
+const ECONNREFUSED: string = "ECONNREFUSED";
+
+export const GET: RequestHandler = async ({ params, fetch }) => {
 	const provider_slug = params.provider_slug;
 
 	interface Account extends WithId<Document> {
@@ -28,9 +30,13 @@ export const GET: RequestHandler = async ({ params }) => {
 
 		await fetch(url + new URLSearchParams({ accessToken: account.access_token }))
 			.then((res) => res.json())
-			.then((json) => result = json.map((item: any) => item.tradeName).join(','));
+			.then((json) => result = json.map((item: any) => item.tradeName).join(','))
+			.catch(() => result = ECONNREFUSED);
 		await collection.deleteOne({ _id: account._id });
 	}
 
+	if (result === ECONNREFUSED) {
+		throw redirect(302, '/dashboard/scan?' + new URLSearchParams({ success: "false", provider: provider_slug, data: result }));
+	}
 	throw redirect(302, '/dashboard/scan?' + new URLSearchParams({ success: "true", provider: provider_slug, data: result }));
 };
